Project-library/requirements.txt
app==0.0.1
arrow==1.3.0
attrs==25.1.0
binaryornot==0.4.4
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.3
cookiecutter==2.1.1
create-app==0.6.2
cryptography==44.0.0
flasgger==0.9.7.1
Flask==2.3.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jinja2-time==0.2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
PyJWT==2.10.1
MarkupSafe==3.0.2
mistune==3.1.1
packaging==24.2
pycparser==2.22
python-dateutil==2.9.0.post0
python-slugify==8.0.4
psycopg2-binary==2.9.10
PyYAML==6.0.2
referencing==0.36.2
requests==2.28.1
rpds-py==0.22.3
six==1.17.0
SQLAlchemy==2.0.37
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==3.1.3
pymongo==4.3.3
redis==4.5.4
Flask-PyMongo==2.3.0
Flask-Caching==2.0.2